refacto: moved compute_target to core
