+++ test/test_models_recognition.py
@@ -146,41 +146,27 @@ def test_recognition_zoo_error():
         _ = recognition.zoo.recognition_predictor("my_fancy_model", pretrained=False)
 
 
-@pytest.fixture(scope="function")
-def test_compute_target_sar():
+def test_compute_loss_sar():
     list_gts = ['elephants', '1234', 'Rémouleur']
-    model = recognition.sar_vgg16_bn()
-    encoded_gts, seq_len = model.compute_target(list_gts)
-    assert isinstance(encoded_gts, tf.Tensor)
-    assert isinstance(seq_len, tf.Tensor)
-    assert list(seq_len.numpy()) == [9, 4, 9]
-    return encoded_gts, seq_len
-
-
-def test_compute_loss_sar(test_compute_target_sar):
     model_input = tf.random.uniform(shape=[3, 32, 128, 3], minval=0, maxval=1)
     model = recognition.sar_vgg16_bn()
     model_output = model(model_input)
-    encoded_gts, seq_len = test_compute_target_sar
+    encoded_gts, seq_len = model.compute_target(list_gts)
     loss = model.compute_loss(encoded_gts, model_output, seq_len)
     assert isinstance(loss, tf.Tensor)
-
-
-@pytest.fixture(scope="function")
-def test_compute_target_crnn():
-    list_gts = ['elephants', '1234', 'Rémouleur']
-    model = recognition.crnn_vgg16_bn()
-    encoded_gts, seq_len = model.compute_target(list_gts)
     assert isinstance(encoded_gts, tf.Tensor)
     assert isinstance(seq_len, tf.Tensor)
     assert list(seq_len.numpy()) == [9, 4, 9]
-    return encoded_gts, seq_len
 
 
-def test_compute_loss_crnn(test_compute_target_crnn):
+def test_compute_loss_crnn():
+    list_gts = ['elephants', '1234', 'Rémouleur']
     model_input = tf.random.uniform(shape=[3, 32, 128, 3], minval=0, maxval=1)
     model = recognition.crnn_vgg16_bn()
     model_output = model(model_input)
-    encoded_gts, seq_len = test_compute_target_crnn
+    encoded_gts, seq_len = model.compute_target(list_gts)
     loss = model.compute_loss(encoded_gts, model_output, seq_len)
     assert isinstance(loss, tf.Tensor)
+    assert isinstance(encoded_gts, tf.Tensor)
+    assert isinstance(seq_len, tf.Tensor)
+    assert list(seq_len.numpy()) == [9, 4, 9]
